refactor: utils 디렉토리 model 하위로 이동
