+++ src/Model/utils/assessWinner.js
