+++ src/Model/utils/movePosition.js
@@ -1,5 +1,5 @@
 import { Random } from '@woowacourse/mission-utils';
-import { MOVESET, RANDOMSET } from '../Constant/SETTING.js';
+import { MOVESET, RANDOMSET } from '../../Constant/SETTING.js';
 
 export default function movePosition() {
   const number = Random.pickNumberInRange(RANDOMSET.startNum, RANDOMSET.endNum);
